+++ src/core/cli.h
@@ -68,6 +68,7 @@ extern char *cli_uds_file;
 
 //------------------------------------------------------------------------------------
 extern bool dump_to_cli;   // If set, dump logging messages are sent back to the CLI client rather than their normal destination
+extern bool is_running_cli_command; // Set if this executable is running a CLI command
 extern bool is_running_cli_local_command; // Set if this executable is running a local CLI command (eg dbset)
 
 #endif